refactor #222 : formatter취합, useTodo훅스 수정
