+++ src/hooks/useReceipts.jsx
@@ -5,8 +5,24 @@ import {
   updatePinnedReceipt,
 } from "controllers/receiptController";
 import { getUser } from "controllers/userController";
+
+export function useFetchReceipt() {
+  const [pinnedReceipts, setPinnedReceipts] = useState();
+
+  const fetchReceipts = useCallback(async () => {
+    const data = await getPinnedReceipts();
+    setPinnedReceipts(data);
+  }, []);
+
+  useEffect(() => {
+    fetchReceipts();
+  }, [fetchReceipts]);
+
+  return { pinnedReceipts, fetchReceipts };
+}
+
 /**
- * 세션 스토리지와 유저의 정보를 취합합니다.
+ * 세션 스토리지와 유저의 정보를 취합하여 새 핀 영수증을 만듧니다.
  *
  * @returns {Object} Pinned Receipt
  */
@@ -30,6 +46,7 @@ function useGeneratePinnedReceipt() {
 
 export function usePinnedReceipt() {
   const { receiptBody } = useGeneratePinnedReceipt();
+  // 여기에 아이디 저장한다고 업데이트가 정상적으로 되는지..?
   const [id, setId] = useState("");
 
   // TODO : 에러 처리
@@ -43,19 +60,4 @@ export function usePinnedReceipt() {
   }
 
   return { postPinReceipt, updatePinReceipt };
-}
-
-export function useFetchReceipt() {
-  const [pinnedReceipts, setPinnedReceipts] = useState();
-
-  const fetchReceipts = useCallback(async () => {
-    const data = await getPinnedReceipts();
-    setPinnedReceipts(data);
-  }, []);
-
-  useEffect(() => {
-    fetchReceipts();
-  }, [fetchReceipts]);
-
-  return { pinnedReceipts, fetchReceipts };
 }
